minor: relaxing the version regex
This patch updates the version regex to be a little less restrictive and play nicer with custom build names. For example, I have a custom build for MongoDB '2.4.6-sub' that doesn't match the current regex due to its requirement for 'rc' in the name.

The new regex would match all of the following examples:

2.4.6-sub
2.4.6-rc999
2.4.6.custom
2.4.6-anythinguntilendofline

This will give us a little more flexibility in naming and identifying custom server versions under m.