Default installation prompt to yes.
Apt (the package manager) has had this default for as long as I remember
which is many years, and if it works for an operating system's
package manager it should also work for a Mongo version manager.

Also changed the return value of m when the user answers in the
negative, as the operation fails in that scenario.