m: fix error message when using GNU egrep
Since GNU grep release 3.8 the egrep command prints a warning to STDERR
that 'egrep is obsolescent', which is annoying. This commit change m to
use `grep -E` instead, which is supported by both BSD and GNU grep.

Announcement of change to GNU egrep:
https://git.savannah.gnu.org/cgit/grep.git/tree/NEWS#n75

The GNU egrep warning can be found here:
https://git.savannah.gnu.org/cgit/grep.git/tree/src/egrep.sh#n3